+++ src/core/modelplugin.cpp
@@ -35,20 +35,19 @@ ModelPlugin::ModelPlugin(QPluginLoader* loader, const QString& libPath)
         return;
     }
 
-    QJsonObject metaData = loader->metaData().value("MetaData").toObject();
-    if (!readAttrsScope(&metaData, PLUGIN_ATTR_NODESCOPE, m_nodeAttrsScope, m_nodeAttrNames)) {
+    if (!readAttrsScope(PLUGIN_ATTR_NODESCOPE, m_nodeAttrsScope, m_nodeAttrNames)) {
         qWarning() << "failed to read the node's attributes!";
         m_type = PluginType::Invalid;
         return;
     }
 
-    if (!readAttrsScope(&metaData, PLUGIN_ATTR_EDGESCOPE, m_edgeAttrsScope, m_edgeAttrNames)) {
+    if (!readAttrsScope(PLUGIN_ATTR_EDGESCOPE, m_edgeAttrsScope, m_edgeAttrNames)) {
         qWarning() << "failed to read the edge's attributes!";
         m_type = PluginType::Invalid;
         return;
     }
 
-    QJsonArray supportedGraphs = metaData.value(PLUGIN_ATTR_SUPPORTEDGRAPHS).toArray();
+    QJsonArray supportedGraphs = m_metaData.value(PLUGIN_ATTR_SUPPORTEDGRAPHS).toArray();
     m_supportedGraphs.reserve(supportedGraphs.size());
     for (QJsonValueRef v : supportedGraphs) {
         QString graphId = v.toString();
@@ -57,7 +56,7 @@ ModelPlugin::ModelPlugin(QPluginLoader* loader, const QString& libPath)
         }
     }
 
-    QJsonArray customOutputs = metaData.value(PLUGIN_ATTR_CUSTOMOUTPUTS).toArray();
+    QJsonArray customOutputs = m_metaData.value(PLUGIN_ATTR_CUSTOMOUTPUTS).toArray();
     m_customOutputs.reserve(customOutputs.size());
     for (QJsonValueRef v : customOutputs) {
         QString function = v.toString();